Import sanity
* break out register_exchanges functions into a separate module. This avoids
  cross-module inter dependencies and makes it possible to actually import modules.
* Also, convinced myself that register_exchanges is not a good candidate for pure
  utils since it assume a logger from the bootstrap namespace etc so those mental
  gymnastics helped to rationalize change.
